requirements.txt
flask
tensorflow
h5py
Werkzeug
numpy
Pillow
scikit-learn
transformers
torch
tf-keras